Used python package dis to disassemble the code to locate and remediate where unnecessary code was adding additional execution time in order to optimize the code.
Some functions were assigning data to a variable that was only used to return the same data. By removing those variables and returning the same data saves execution time by not having to allocate memory for a variable.

Simplified some comparisons that did not to be complex. Such as comparisons that were checking whether something was True or not. This is already done by default when just calling the variable inside of a comparison, when adding the == True when not needed it adds additional slowdowns.